Re-implement template functionality
